# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "mousecake-go 区块链后端服务 API 文档，包含用户认证、报价聚合等模块。",
        "title": "mousecake-go API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/admin/login": {
            "post": {
                "description": "使用用户名和密码登录管理员账号，签发 JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "管理员登录",
                "parameters": [
                    {
                        "description": "管理员登录请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_user.adminLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_user.tokenResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据 Bearer Token 获取当前登录用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "获取当前用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_user.userInfoResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            }
        },
        "/auth/wallet/nonce": {
            "post": {
                "description": "根据钱包地址生成 SIWE 登录消息和 nonce",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "获取钱包登录 nonce",
                "parameters": [
                    {
                        "description": "钱包地址",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_user.walletNonceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_user.nonceResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            }
        },
        "/auth/wallet/verify": {
            "post": {
                "description": "验证 SIWE 消息签名，签发 JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "验证钱包签名",
                "parameters": [
                    {
                        "description": "签名验证请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_user.walletVerifyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_user.tokenResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code 业务状态码，0 表示成功，非 0 表示失败。",
                    "type": "integer"
                },
                "data": {
                    "description": "Data 响应数据载荷，成功时为业务数据，失败时为 nil。"
                },
                "message": {
                    "description": "Message 响应消息描述。",
                    "type": "string"
                }
            }
        },
...
        "internal_user.adminLoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "********"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "internal_user.nonceResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "mousecake-go wants you to sign in..."
                },
                "nonce": {
                    "type": "string",
                    "example": "abc123def456"
                }
            }
        },
        "internal_user.tokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIs..."
                },
                "expires_in": {
                    "type": "integer",
                    "example": 900
                },
                "token_type": {
                    "type": "string",
                    "example": "Bearer"
                }
            }
        },
        "internal_user.userInfoResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "0x1234567890abcdef1234567890abcdef12345678"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_admin": {
                    "type": "boolean",
                    "example": false
                },
                "name": {
                    "type": "string",
                    "example": "张三"
                },
                "nickname": {
                    "type": "string",
                    "example": "mouse"
                },
                "type": {
                    "type": "string",
                    "example": "wallet"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "internal_user.walletNonceRequest": {
            "type": "object",
            "required": [
                "address"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "example": "0x1234567890abcdef1234567890abcdef12345678"
                }
            }
        },
        "internal_user.walletVerifyRequest": {
            "type": "object",
            "required": [
                "message",
                "signature"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "example": "mousecake-go wants you to sign in with your Ethereum account..."
                },
                "signature": {
                    "type": "string",
                    "example": "0xabcdef..."
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}